Refactor bootstrap clients (#8261)
* add bootstrap/clients pkg

* use new bootstrap/clients

* transparently use bootstrap/clients from bootstrap

* remove stray test file

* add changelog

* K8s tests in kind (#8235)

* move kubes tests to new folder

* add kube package to ci regression tests action

* empty

* move tests to kube2e/gloo

* modify test to use real kube client

* codegen

* cleanup, references to original locations

* changelog

* remove RUN_KUBE_TESTS from env.go

* move changelog

* remove unneccesary before and after suite steps

* temp fix for breaking change

* remove unused import

* dont use utils name

* focus aws test without k8s client

* move happy path kubernetes tests to kube2e

* unfocus, remove k8s client completely

* focus ec2, comment kube client

* remove kube client

* remove kubernetes requirement

* formatting

* remove cluster lock

* cleanup

* remove unneeded crd code

* remove deprecated helper

* remove comment

* remove deprecated func

---------


* remove kube tests from new location

* add/correct godoc comments

* fix wrapped aliased type, make imports more idiomatic

* redirect deprecated consul and vault functinos

* Update projects/gloo/pkg/bootstrap/clients/consul.go

---------
