+++ changelog/v1.15.0-beta8/refactor-bootstrap.yaml
@@ -1,0 +1,5 @@
+changelog:
+- type: NON_USER_FACING
+  description: >-
+    Refactor bootstrap package to add clarifying distinction between opts and clients. Additionally, breaks
+    clients into their own files to make it easier to work with.
