+++ projects/gloo/pkg/bootstrap/utils.go
@@ -2,171 +2,69 @@ package bootstrap
 
 import (
 	"context"
-	"path/filepath"
-	"time"
 
-	"github.com/golang/protobuf/ptypes/duration"
-	"github.com/solo-io/gloo/projects/gloo/pkg/defaults"
 	"github.com/solo-io/solo-kit/pkg/api/v1/clients/kube/cache"
-	"github.com/solo-io/solo-kit/pkg/utils/prototime"
 
 	consulapi "github.com/hashicorp/consul/api"
 	vaultapi "github.com/hashicorp/vault/api"
-	kubeconverters "github.com/solo-io/gloo/projects/gloo/pkg/api/converters/kube"
 	v1 "github.com/solo-io/gloo/projects/gloo/pkg/api/v1"
-	"github.com/solo-io/k8s-utils/kubeutils"
-	"github.com/solo-io/solo-kit/pkg/api/external/kubernetes/service"
+	"github.com/solo-io/gloo/projects/gloo/pkg/bootstrap/clients"
+	"github.com/solo-io/gloo/projects/gloo/pkg/defaults"
 	"github.com/solo-io/solo-kit/pkg/api/v1/clients/factory"
 	"github.com/solo-io/solo-kit/pkg/api/v1/clients/kube"
 	"github.com/solo-io/solo-kit/pkg/api/v1/clients/kube/crd"
 	"github.com/solo-io/solo-kit/pkg/api/v1/clients/memory"
 	skkube "github.com/solo-io/solo-kit/pkg/api/v1/resources/common/kubernetes"
-	"github.com/solo-io/solo-kit/pkg/errors"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/rest"
 )
 
 const (
-	DefaultK8sQPS   = 50     // 10x the k8s-recommended default; gloo gets busy writing status updates
-	DefaultK8sBurst = 100    // 10x the k8s-recommended default; gloo gets busy writing status updates
-	DefaultRootKey  = "gloo" // used for vault and consul key-value storage
+	// Deprecated. Use bootstrap/clients.
+	DefaultK8sQPS = clients.DefaultK8sQPS // 10x the k8s-recommended default; gloo gets busy writing status updates
+	// Deprecated. Use bootstrap/clients.
+	DefaultK8sBurst = clients.DefaultK8sBurst // 10x the k8s-recommended default; gloo gets busy writing status updates
+	// Deprecated. Use bootstrap/clients.
+	DefaultRootKey = clients.DefaultRootKey // used for vault and consul key-value storage
 )
 
-// default query options for consul
-var DefaultQueryOptions = &consulapi.QueryOptions{RequireConsistent: true, AllowStale: false}
+// Deprecated. Use bootstrap/clients.
+var DefaultQueryOptions = clients.DefaultConsulQueryOptions
 
-type ConfigFactoryParams struct {
-	settings *v1.Settings
-	memory   configFactoryParamsMemory
-	kube     configFactoryParamsKube
-	consul   configFactoryParamsConsul
-}
+// Deprecated. Use bootstrap/clients.
+type ConfigFactoryParams = clients.ConfigFactoryParams
 
+// Deprecated. Use bootstrap/clients.
 func NewConfigFactoryParams(settings *v1.Settings,
 	sharedCache memory.InMemoryResourceCache,
 	cache kube.SharedCache,
 	cfg **rest.Config,
 	consulClient *consulapi.Client) ConfigFactoryParams {
-	return ConfigFactoryParams{
-		settings: settings,
-		memory: configFactoryParamsMemory{
-			sharedCache: sharedCache,
-		},
-		kube: configFactoryParamsKube{
-			kubeCache: cache,
-			restCfg:   cfg,
-		},
-		consul: configFactoryParamsConsul{
-			consulClient: consulClient,
-		},
-	}
+	return clients.NewConfigFactoryParams(settings, sharedCache, cache, cfg, consulClient)
 }
 
-type configFactoryParamsMemory struct {
-	sharedCache memory.InMemoryResourceCache
-}
-
-type configFactoryParamsKube struct {
-	kubeCache kube.SharedCache
-	restCfg   **rest.Config
-}
-
-type configFactoryParamsConsul struct {
-	consulClient *consulapi.Client
-}
-
-// sharedCache, resourceCrd+cfg OR consulClient must be non-nil
+// Deprecated. Use bootstrap/clients.
 func ConfigFactoryForSettings(params ConfigFactoryParams, resourceCrd crd.Crd) (factory.ResourceClientFactory, error) {
-	settings := params.settings
-
-	if settings.GetConfigSource() == nil {
-		sharedCache := params.memory.sharedCache
-		if sharedCache == nil {
-			return nil, errors.Errorf("internal error: shared cache cannot be nil")
-		}
-		return &factory.MemoryResourceClientFactory{
-			Cache: sharedCache,
-		}, nil
-	}
-
-	switch source := settings.GetConfigSource().(type) {
-	// this is at trick to reuse the same cfg across multiple clients
-	case *v1.Settings_KubernetesConfigSource:
-		kubeCache := params.kube.kubeCache
-		cfg := params.kube.restCfg
-		if *cfg == nil {
-			c, err := kubeutils.GetConfig("", "")
-			if err != nil {
-				return nil, err
-			}
-
-			c.QPS = DefaultK8sQPS
-			c.Burst = DefaultK8sBurst
-			if kubeSettingsConfig := settings.GetKubernetes(); kubeSettingsConfig != nil {
-				if rl := kubeSettingsConfig.GetRateLimits(); rl != nil {
-					c.QPS = rl.GetQPS()
-					c.Burst = int(rl.GetBurst())
-				}
-			}
-			*cfg = c
-		}
-
-		return &factory.KubeResourceClientFactory{
-			Crd:                resourceCrd,
-			Cfg:                *cfg,
-			SharedCache:        kubeCache,
-			NamespaceWhitelist: settings.GetWatchNamespaces(),
-		}, nil
-	case *v1.Settings_ConsulKvSource:
-		consulClient := params.consul.consulClient
-		rootKey := source.ConsulKvSource.GetRootKey()
-		if rootKey == "" {
-			rootKey = DefaultRootKey
-		}
-		return &factory.ConsulResourceClientFactory{
-			Consul:       consulClient,
-			RootKey:      rootKey,
-			QueryOptions: DefaultQueryOptions,
-		}, nil
-	case *v1.Settings_DirectoryConfigSource:
-		return &factory.FileResourceClientFactory{
-			RootDir: filepath.Join(source.DirectoryConfigSource.GetDirectory(), resourceCrd.Plural),
-		}, nil
-	}
-	return nil, errors.Errorf("invalid config source type")
+	return clients.ConfigFactoryForSettings(params, resourceCrd)
 }
 
+// Deprecated. Use bootstrap/clients.
 func KubeServiceClientForSettings(ctx context.Context,
 	settings *v1.Settings,
 	sharedCache memory.InMemoryResourceCache,
 	cfg **rest.Config,
 	clientset *kubernetes.Interface,
 	kubeCoreCache *cache.KubeCoreCache) (skkube.ServiceClient, error) {
-
-	// We are running in kubernetes
-	switch settings.GetConfigSource().(type) {
-	case *v1.Settings_KubernetesConfigSource:
-		if err := initializeForKube(ctx, cfg, clientset, kubeCoreCache, settings.GetRefreshRate(), settings.GetWatchNamespaces()); err != nil {
-			return nil, errors.Wrapf(err, "initializing kube cfg clientset and core cache")
-		}
-		return service.NewServiceClient(*clientset, *kubeCoreCache), nil
-	}
-
-	// In all other cases, run in memory
-	if sharedCache == nil {
-		return nil, errors.Errorf("internal error: shared cache cannot be nil")
-	}
-	memoryRcFactory := &factory.MemoryResourceClientFactory{Cache: sharedCache}
-	inMemoryClient, err := memoryRcFactory.NewResourceClient(ctx, factory.NewResourceClientParams{
-		ResourceType: &skkube.Service{},
-	})
-	if err != nil {
-		return nil, err
-	}
-	return skkube.NewServiceClientWithBase(inMemoryClient), nil
+	return clients.KubeServiceClientForSettings(ctx,
+		settings,
+		sharedCache,
+		cfg,
+		clientset,
+		kubeCoreCache,
+	)
 }
 
-// sharedCache OR resourceCrd+cfg must be non-nil
+// Deprecated. Use bootstrap/clients.
 func SecretFactoryForSettings(ctx context.Context,
 	settings *v1.Settings,
 	sharedCache memory.InMemoryResourceCache,
@@ -175,44 +73,18 @@ func SecretFactoryForSettings(ctx context.Context,
 	kubeCoreCache *cache.KubeCoreCache,
 	vaultClient *vaultapi.Client,
 	pluralName string) (factory.ResourceClientFactory, error) {
-	if settings.GetSecretSource() == nil {
-		if sharedCache == nil {
-			return nil, errors.Errorf("internal error: shared cache cannot be nil")
-		}
-		return &factory.MemoryResourceClientFactory{
-			Cache: sharedCache,
-		}, nil
-	}
-
-	switch source := settings.GetSecretSource().(type) {
-	case *v1.Settings_KubernetesSecretSource:
-		if err := initializeForKube(ctx, cfg, clientset, kubeCoreCache, settings.GetRefreshRate(), settings.GetWatchNamespaces()); err != nil {
-			return nil, errors.Wrapf(err, "initializing kube cfg clientset and core cache")
-		}
-		return &factory.KubeSecretClientFactory{
-			Clientset:       *clientset,
-			Cache:           *kubeCoreCache,
-			SecretConverter: kubeconverters.GlooSecretConverterChain,
-		}, nil
-	case *v1.Settings_VaultSecretSource:
-		rootKey := source.VaultSecretSource.GetRootKey()
-		if rootKey == "" {
-			rootKey = DefaultRootKey
-		}
-		pathPrefix := source.VaultSecretSource.GetPathPrefix()
-		if pathPrefix == "" {
-			pathPrefix = DefaultPathPrefix
-		}
-		return NewVaultSecretClientFactory(vaultClient, pathPrefix, rootKey), nil
-	case *v1.Settings_DirectorySecretSource:
-		return &factory.FileResourceClientFactory{
-			RootDir: filepath.Join(source.DirectorySecretSource.GetDirectory(), pluralName),
-		}, nil
-	}
-	return nil, errors.Errorf("invalid config source type")
+	return clients.SecretFactoryForSettings(ctx,
+		settings,
+		sharedCache,
+		cfg,
+		clientset,
+		kubeCoreCache,
+		vaultClient,
+		pluralName,
+	)
 }
 
-// sharedCach OR resourceCrd+cfg must be non-nil
+// Deprecated. Use bootstrap/clients.
 func ArtifactFactoryForSettings(ctx context.Context,
 	settings *v1.Settings,
 	sharedCache memory.InMemoryResourceCache,
@@ -221,82 +93,19 @@ func ArtifactFactoryForSettings(ctx context.Context,
 	kubeCoreCache *cache.KubeCoreCache,
 	consulClient *consulapi.Client,
 	pluralName string) (factory.ResourceClientFactory, error) {
-	if settings.GetArtifactSource() == nil {
-		if sharedCache == nil {
-			return nil, errors.Errorf("internal error: shared cache cannot be nil")
-		}
-		return &factory.MemoryResourceClientFactory{
-			Cache: sharedCache,
-		}, nil
-	}
-
-	switch source := settings.GetArtifactSource().(type) {
-	case *v1.Settings_KubernetesArtifactSource:
-		if err := initializeForKube(ctx, cfg, clientset, kubeCoreCache, settings.GetRefreshRate(), settings.GetWatchNamespaces()); err != nil {
-			return nil, errors.Wrapf(err, "initializing kube cfg clientset and core cache")
-		}
-		return &factory.KubeConfigMapClientFactory{
-			Clientset:       *clientset,
-			Cache:           *kubeCoreCache,
-			CustomConverter: kubeconverters.NewArtifactConverter(),
-		}, nil
-	case *v1.Settings_DirectoryArtifactSource:
-		return &factory.FileResourceClientFactory{
-			RootDir: filepath.Join(source.DirectoryArtifactSource.GetDirectory(), pluralName),
-		}, nil
-	case *v1.Settings_ConsulKvArtifactSource:
-		rootKey := source.ConsulKvArtifactSource.GetRootKey()
-		if rootKey == "" {
-			rootKey = DefaultRootKey
-		}
-		return &factory.ConsulResourceClientFactory{
-			Consul:       consulClient,
-			RootKey:      rootKey,
-			QueryOptions: DefaultQueryOptions,
-		}, nil
-	}
-	return nil, errors.Errorf("invalid config source type")
+	return clients.ArtifactFactoryForSettings(ctx,
+		settings,
+		sharedCache,
+		cfg,
+		clientset,
+		kubeCoreCache,
+		consulClient,
+		pluralName,
+	)
 }
 
-func initializeForKube(ctx context.Context,
-	cfg **rest.Config,
-	clientset *kubernetes.Interface,
-	kubeCoreCache *cache.KubeCoreCache,
-	refreshRate *duration.Duration,
-	nsToWatch []string,
-) error {
-	if cfg == nil {
-		c, err := kubeutils.GetConfig("", "")
-		if err != nil {
-			return err
-		}
-		*cfg = c
-	}
-
-	if *clientset == nil {
-		cs, err := kubernetes.NewForConfig(*cfg)
-		if err != nil {
-			return err
-		}
-		*clientset = cs
-	}
-
-	if *kubeCoreCache == nil {
-		duration := 12 * time.Hour
-		if refreshRate != nil {
-			duration = prototime.DurationFromProto(refreshRate)
-		}
-		coreCache, err := cache.NewKubeCoreCacheWithOptions(ctx, *clientset, duration, nsToWatch)
-		if err != nil {
-			return err
-		}
-		*kubeCoreCache = coreCache
-	}
-
-	return nil
-
-}
-
+// GetWriteNamespace checks the provided settings for the field `DiscoveryNamespace`
+// and defaults to `defaults.GlooSystem` if not found.
 func GetWriteNamespace(settings *v1.Settings) string {
 	writeNamespace := settings.GetDiscoveryNamespace()
 	if writeNamespace == "" {
